Rename `classMainWindow.py` -> `main_window.py`.
